add default query endpoint for project
