+++ lib/api/v3/queries/queries_api.rb
@@ -42,6 +42,8 @@ module API
           mount API::V3::Queries::Schemas::QuerySchemaAPI
           mount API::V3::Queries::Schemas::QueryFilterInstanceSchemaAPI
 
+          helpers ::API::V3::Queries::Helpers::QueryRepresenterResponse
+
           helpers do
             def authorize_by_policy(action, &block)
               authorize_by_with_raise(-> () { allowed_to?(action) }, &block)
@@ -49,21 +51,6 @@ module API
 
             def allowed_to?(action)
               QueryPolicy.new(current_user).allowed?(@query, action)
-            end
-
-            def query_representer(query, params)
-              representer = ::API::V3::WorkPackageCollectionFromQueryService
-                            .new(query, current_user)
-                            .call(params)
-
-              if representer.success?
-                QueryRepresenter.new(query,
-                                     current_user: current_user,
-                                     results: representer.result,
-                                     params: params)
-              else
-                raise ::API::Errors::InvalidQuery.new(representer.errors.full_messages)
-              end
             end
           end
 
@@ -97,7 +84,7 @@ module API
 
               authorize_by_policy(:show)
 
-              query_representer(@query, params)
+              query_representer_response(@query, params)
             end
           end
 
@@ -114,7 +101,7 @@ module API
             end
 
             get do
-              query_representer(@query, params)
+              query_representer_response(@query, params)
             end
 
             delete do
@@ -138,13 +125,13 @@ module API
               end
               query_menu_item.save!
 
-              query_representer(@query, {})
+              query_representer_response(@query, {})
             end
 
             patch :unstar do
               authorize_by_policy(:unstar)
 
-              representer = query_representer(@query, {})
+              representer = query_representer_response(@query, {})
 
               query_menu_item = @query.query_menu_item
               return representer if @query.query_menu_item.nil?
